(#148) Delegated the validation against json schemas to dedicated class
